+++ xmake/platforms/windows/config.lua
@@ -125,7 +125,7 @@ function main(platform, name)
 
     -- only check the given config name?
     if name then
-        local toolchain = singleton.get("windows.toolchains", _toolchains)[name]
+        local toolchain = singleton.get("windows.toolchains." .. (config.get("arch") or os.arch()), _toolchains)[name]
         if toolchain then
             environment.enter("toolchains")
             check_toolchain(config, name, toolchain)